fix: Fix encoder panic and encoding bugs, overhaul test suite
Bugs fixed:
- custom levels outside [-2, 7] crashed the encoder (colour table
  indexed by raw level); unknown levels now render panic-style
- zap.Durations doubled the list separator; zap.Times additionally
  escaped the time encoder's own ANSI codes into literal text
- ByteString fields added via Logger.With rendered base64
- hand-rolled EncoderConfigs panicked on nil EncodeDuration/EncodeTime/
  NewReflectedEncoder; empty LineEnding disabled namespace layout;
  SkipLineEnding was ignored

Test suite:
- expectations move to tagged golden files (testdata/*.golden,
  regenerated with go test -update); ANSI codes shown as readable tags
- test files now mirror source files; escaping code split into
  escape.go with matching escape_test.go
- new coverage (77.9% -> 96.6%): escaping perturbations, ANSI injection,
  every recording/object/array encoder method, callback fallbacks,
  levels, caller, zero-value configs, config knobs, sugared and
  production-style logger usage, fuzzing, concurrency under -race,
  encoder reuse and clone independence
