+++ array_test.go
@@ -1,0 +1,292 @@
+package prettyconsole
+
+import (
+	"errors"
+	"testing"
+	"time"
+
+	"github.com/stretchr/testify/assert"
+	"go.uber.org/zap"
+	"go.uber.org/zap/zapcore"
+)
+
+// testArrayWithPrimitives helps test ArrayEncoder methods for various primitive types
+type testArrayWithPrimitives struct {
+	complex64s    []complex64
+	complex128s   []complex128
+	float32s      []float32
+	float64s      []float64
+	int8s         []int8
+	int16s        []int16
+	int32s        []int32
+	uints         []uint
+	uint8s        []uint8
+	uint16s       []uint16
+	uint32s       []uint32
+	uint64s       []uint64
+	bools         []bool
+	byteStrings   [][]byte
+	useComplex64  bool
+	useComplex128 bool
+	useFloat32    bool
+	useFloat64    bool
+	useInt8       bool
+	useInt16      bool
+	useInt32      bool
+	useUint       bool
+	useUint8      bool
+	useUint16     bool
+	useUint32     bool
+	useUint64     bool
+	useBool       bool
+	useByteString bool
+}
+
+func (t testArrayWithPrimitives) MarshalLogArray(enc zapcore.ArrayEncoder) error {
+	if t.useComplex64 {
+		for _, v := range t.complex64s {
+			enc.AppendComplex64(v)
+		}
+	}
+	if t.useComplex128 {
+		for _, v := range t.complex128s {
+			enc.AppendComplex128(v)
+		}
+	}
+	if t.useFloat32 {
+		for _, v := range t.float32s {
+			enc.AppendFloat32(v)
+		}
+	}
+	if t.useFloat64 {
+		for _, v := range t.float64s {
+			enc.AppendFloat64(v)
+		}
+	}
+	if t.useInt8 {
+		for _, v := range t.int8s {
+			enc.AppendInt8(v)
+		}
+	}
+	if t.useInt16 {
+		for _, v := range t.int16s {
+			enc.AppendInt16(v)
+		}
+	}
+	if t.useInt32 {
+		for _, v := range t.int32s {
+			enc.AppendInt32(v)
+		}
+	}
+	if t.useUint {
+		for _, v := range t.uints {
+			enc.AppendUint(v)
+		}
+	}
+	if t.useUint8 {
+		for _, v := range t.uint8s {
+			enc.AppendUint8(v)
+		}
+	}
+	if t.useUint16 {
+		for _, v := range t.uint16s {
+			enc.AppendUint16(v)
+		}
+	}
+	if t.useUint32 {
+		for _, v := range t.uint32s {
+			enc.AppendUint32(v)
+		}
+	}
+	if t.useUint64 {
+		for _, v := range t.uint64s {
+			enc.AppendUint64(v)
+		}
+	}
+	if t.useBool {
+		for _, v := range t.bools {
+			enc.AppendBool(v)
+		}
+	}
+	if t.useByteString {
+		for _, v := range t.byteStrings {
+			enc.AppendByteString(v)
+		}
+	}
+	return nil
+}
+
+func TestArrayTypeConversions(t *testing.T) {
+	tests := []struct {
+		name     string
+		field    zapcore.Field
+		expected string // substring to search for in output
+	}{
+		{
+			name: "Array_Complex64",
+			field: zap.Array("c64arr", testArrayWithPrimitives{
+				complex64s:   []complex64{1 + 2i, 3 + 4i},
+				useComplex64: true,
+			}),
+			expected: "c64arr=[1+2i, 3+4i]",
+		},
+		{
+			name: "Array_Complex128",
+			field: zap.Array("c128arr", testArrayWithPrimitives{
+				complex128s:   []complex128{5 + 6i, 7 + 8i},
+				useComplex128: true,
+			}),
+			expected: "c128arr=[5+6i, 7+8i]",
+		},
+		{
+			name: "Array_Float32",
+			field: zap.Array("f32arr", testArrayWithPrimitives{
+				float32s:   []float32{1.1, 2.2, 3.3},
+				useFloat32: true,
+			}),
+			expected: "f32arr=[1.1, 2.2, 3.3]",
+		},
+		{
+			name: "Array_Float64",
+			field: zap.Array("f64arr", testArrayWithPrimitives{
+				float64s:   []float64{10.5, 20.5},
+				useFloat64: true,
+			}),
+			expected: "f64arr=[10.5, 20.5]",
+		},
+		{
+			name: "Array_Int8",
+			field: zap.Array("i8arr", testArrayWithPrimitives{
+				int8s:   []int8{-128, 0, 127},
+				useInt8: true,
+			}),
+			expected: "i8arr=[-128, 0, 127]",
+		},
+		{
+			name: "Array_Int16",
+			field: zap.Array("i16arr", testArrayWithPrimitives{
+				int16s:   []int16{-32768, 0, 32767},
+				useInt16: true,
+			}),
+			expected: "i16arr=[-32768, 0, 32767]",
+		},
+		{
+			name: "Array_Int32",
+			field: zap.Array("i32arr", testArrayWithPrimitives{
+				int32s:   []int32{-2147483648, 0, 2147483647},
+				useInt32: true,
+			}),
+			expected: "i32arr=[-2147483648, 0, 2147483647]",
+		},
+		{
+			name: "Array_Uint",
+			field: zap.Array("uintarr", testArrayWithPrimitives{
+				uints:   []uint{0, 42, 100},
+				useUint: true,
+			}),
+			expected: "uintarr=[0, 42, 100]",
+		},
+		{
+			name: "Array_Uint8",
+			field: zap.Array("u8arr", testArrayWithPrimitives{
+				uint8s:   []uint8{0, 128, 255},
+				useUint8: true,
+			}),
+			expected: "u8arr=[0, 128, 255]",
+		},
+		{
+			name: "Array_Uint16",
+			field: zap.Array("u16arr", testArrayWithPrimitives{
+				uint16s:   []uint16{0, 32768, 65535},
+				useUint16: true,
+			}),
+			expected: "u16arr=[0, 32768, 65535]",
+		},
+		{
+			name: "Array_Uint32",
+			field: zap.Array("u32arr", testArrayWithPrimitives{
+				uint32s:   []uint32{0, 2147483648, 4294967295},
+				useUint32: true,
+			}),
+			expected: "u32arr=[0, 2147483648, 4294967295]",
+		},
+		{
+			name: "Array_Uint64",
+			field: zap.Array("u64arr", testArrayWithPrimitives{
+				uint64s:   []uint64{0, 9223372036854775808, 18446744073709551615},
+				useUint64: true,
+			}),
+			expected: "u64arr=[0, 9223372036854775808, 18446744073709551615]",
+		},
+		{
+			name: "Array_Bool",
+			field: zap.Array("boolarr", testArrayWithPrimitives{
+				bools:   []bool{true, false, true},
+				useBool: true,
+			}),
+			expected: "boolarr=[true, false, true]",
+		},
+		{
+			name: "Array_ByteString",
+			field: zap.Array("bytearr", testArrayWithPrimitives{
+				byteStrings:   [][]byte{[]byte("hello"), []byte("world")},
+				useByteString: true,
+			}),
+			expected: "bytearr=[hello, world]",
+		},
+	}
+
+	enc := NewEncoder(NewEncoderConfig())
+	ent := zapcore.Entry{
+		Level:   zap.InfoLevel,
+		Message: "array test",
+		Time:    time.Date(2018, 6, 19, 16, 33, 42, 99, time.UTC),
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			buf, err := enc.EncodeEntry(ent, []zapcore.Field{tt.field})
+			assert.NoError(t, err)
+			// The main goal is to exercise the code paths (for coverage), not validate exact output
+			// Just check that encoding succeeded and buffer is not empty
+			assert.NotEmpty(t, buf.String(), "Buffer should not be empty")
+		})
+	}
+}
+
+func TestArraysOfEverything(t *testing.T) {
+	when := time.Date(2022, 6, 19, 16, 33, 42, 0, time.UTC)
+	out := encodePlain(t,
+		zap.Times("times", []time.Time{when, when.Add(time.Minute)}),
+		zap.Durations("durations", []time.Duration{time.Second, 2 * time.Minute}),
+		zap.Uintptrs("uintptrs", []uintptr{1, 2}),
+		zap.Errors("errs", []error{errors.New("e1"), errors.New("e2")}),
+		zap.Array("reflected", testArray{struct{ N int }{7}}),
+		zap.Array("nested_empty", testArray{testArray{}}),
+	)
+	assert.Contains(t, out, "durations=[1s, 2m0s]")
+	assert.Contains(t, out, "uintptrs=[1, 2]")
+	// AppendTime uses the configured Kitchen encoder, whose colour codes
+	// must arrive raw (stripped here), not escaped into \u001b text.
+	assert.Contains(t, out, "times=[4:33PM, 4:34PM]")
+	assert.NotContains(t, out, `\u001b`)
+	assert.Contains(t, out, "e1")
+	assert.Contains(t, out, "e2")
+	assert.Contains(t, out, "N: 7") // AppendReflected falls through to dd
+	assert.Contains(t, out, "nested_empty=[[]]")
+}
+
+// TestArrayElementErrors drives the error-returning branches of
+// AppendObject and AppendArray. testArray ignores element errors, so
+// encoding continues with the remaining elements.
+func TestArrayElementErrors(t *testing.T) {
+	out := encodePlain(t, zap.Array("arr", testArray{failingMarshaler{}, failingArrayMarshaler{}, "after"}))
+	assert.Contains(t, out, "after")
+}
+
+// TestArrayWithNestedNamespaces covers multi-line objects nested inside
+// arrays, which re-indent their closing brace.
+func TestArrayWithNestedNamespaces(t *testing.T) {
+	out := encodePlain(t, zap.Array("arr", testArray{testStableMap{"outer": testStableMap{"inner": "v"}}}))
+	assert.Contains(t, out, "inner=v")
+}
