+++ testdata/TestEncodeEntryErrors_Minimal_Error.golden
@@ -1,0 +1,2 @@
+<gray>4:33PM<r><red> <r><red>ERR<r><red> <r><bold><red>><r><r><red> <r>test message
+<red>  ↳ error<r><red>=<r>Something <red>\n<r>went wrong
