+++ testdata/TestEncodeEntry_Minimal.golden
@@ -1,0 +1,1 @@
+<gray>4:33PM<r><green> <r><green>INF<r><green> <r><bold><green>><r><r>
